refactoring strategy folder based on strategy pattern
